Fix: a and c matrices must have at least one column in td04ad_r
This allows conversion of static gain (i.e., constant) transfer
functions.

SLICOT and Fortran can possibly handle the 0 case, but f2py seems to not
like dimensions of length 0.

td04ad_c needs a similar fix, but I couldn't get it to work; I suspect
this is due to a bug in SLICOT.