add test for overidden XERBLA
We call the _wrapper function directly so the python wrapper
in analysis.py does not raise an exception or in a future version does
not eventually handle the wrong parameter by checking it before the
Fortran call.

Because the file descriptors for the Fortran module are set at import
time, we cannot use the pytest fixtures capsys or capfd, but need to
start a new process.